Fixed a variable screwup in TagValidCPVs
I had targeted the ID of the instruction to be ?isnt instead of ?inst. This
will cause a new variable to bound to the corresponding instruction. This
change also ensures that correct instructions are selected for the target block
on the wavefront. Fortunately I never stumbled over this too badly but I may
have inadvertantly fixed something that would popup later on.