Fixed the stages prior to pathing to support loops and cycles
It turns out that the issue stems from the fact that the parents of regions
were not being updated properly. This has been fixed by adding in rules that
assert facts that describe what the original parent of each region (including
loops) and during fixup rename will assert a fact describing the changes
necessary in parentage. The update will only occur if the region to be modified
is a region and not a loop.

A new stage will have to eventually be added which goes through after fixup is
completed and remove any elements from the Contents multislot of each region
which are not owned by that region.

To do this I had to add two new stages: FixupRegions (to correct the parentage)
and FixupProducers (to propagate region producers and also identify the set
of non local dependencies for each block).

Currently the code is still broken with respect to ordering loops and nested
loops.