Modified the set of rules in BlockDependencyCreation.clp
1) IdentifyRAW, IdentifyWAW, and IdentifyWAR have been replaced with a single
rule that iterates through the operands of a given instruction and creates
dependency information from that. This change alone significantly reduces the
running time of the scheduler. The reason for this speed boost is that the
previous method of dependency analysis was redundant. The set of operands for a
given instruction defines the exact set of dependencies that instruction has
scheduling wise.

2) Modified the rules having to do with call analysis. The rules for finding
instructions that occur _before_ a given call have been modified to actually be
_more_ procedural but also far more intelligent match wise as well. Instead of
looking through the set of call instructions (expensive) the rule first
captures a basic block (cheap) and iterates through the Contents until a call
instruction is found. The set of elements that occur before the call are
captured and iterated through in the body of the rule.

Both rules for pre-call dependencies have been fixed this way.

The rules for identifying instructions that follow a call have been modified to
use the same iterative method as the pre-call rules. The difference between the
two sets, beyond a logical understanding, is the fact that the post-call
rules do _not_ assert consumption facts for instructions. It was found that
this was unnecessary because the instruction is never going to be scheduled in
the current incarnation of the scheduler.

3) The first set of setify rules has been removed. Performing the operation
twice was found to be unnecessary. It would slow things down greatly.

All of these changes have dropped the running time on my test machine by 0.4
seconds on a small case. Originally the scheduler took 1.6 seconds on this test
code and now it only takes 1.2 seconds.

Also, it is important to note that the original versions of the rules are still
in the code because I may need to use them as reference in the future.
Obviously, they are commented out.