Removed references to the SourceRegisters multislot
This is a duplicate of the Operands field and was causing a large amount of
duplication to occur. Consequently, this has been removed and all references to
the code has been replaced with references to Operands instead. Since
SourceRegisters and Operands contain the same data it should work the same.
The other change made is that the set of producers is no longer populated ahead
of time during object initialization. This was causing an unnecessary amount of
setification and as such has been removed. On the unoptimized version of flops
the running time dropped from 1.22 to 1.07 seconds!

Memory consumption should have also dropped due to the removal of the
unnecessary SourceRegisters multislot attached to each copy of the instructions
in a given function. However, this was not actually looked at during testing.