Rewrote the loop region merging code
While I still use the concept of determining ownership using flat lists, I then
use the facts asserted to figure out which claims are indirect and direct.

Direct claims are claims by regions that do not claim another region claiming
the current region. So if A owns B and B owns C then A owns C indirectly.

It also is necessary to figure out if there are indirect indirect claims. If it
turns out that a claim is made for a given region that is an indirect claim of
another claimer then the region that makes the indirect claim.

It's easier to just look at the rules as they are really simple. I also added
four more stages to handle the ownership determination.