Add support for determining basicblock ownership.
This ownership determination uses the same infrastructure used for loops and
regions. In fact, the only major change is that there are new rules for
creating ownership determinants for basic blocks as well as setting the parent
of a basic block. This fixes the issue of incorrect ownership of basic blocks.

I also added a rule to go through and strip out blocks, regions, and loops that
are not owned by the given region.