Made many changes
1) Split the analyze stage into three parts GenerateAnalyze0, GenerateAnalyze,
   and Analyze. The division ensures that all CPV check facts are processed
   before any instructions are analyzed
2) Fixed a bug where phi nodes were disabling instructions from being scheduled
even if the parents of the phi node and instruction were not the same. This has
been fixed to only disable instructions if the dependent phi has the same
parent block.
3) Fixed subset checks to make sure that nonlocals were searched for in the
scheduled list instead of the instruction list. This is a side effect of the
changes made to preschedule non-locals

4) Added some comments to RemoveScheduleStyleForCPV relating to handling loops
and critical paths within those loops.