Updated the deletion mechanisms for flatlists and OwnershipDeterminants
Originally I had a rule to delete each flatlist and OwnershipDeterminant
individually. This is extremely slow because the agenda must be recomputed each
time an object is deleted. To counteract this, I have set it up that the flat
list deletion rule and the OwnershipDeterminant deletion rule delete _all_ of
the objects in a single rule fire.