Updated functions in PathUpdate to be more efficient
The match pattern $?a ?b $?c is expensive because of the $?a multifield.
It makes it simple to do deletes but also makes the code far more inefficient
when using modify-instance due to the messages being passed.
DetermineIndirectClaim and DetermineIndirectIndirectClaim have been modified to
use the $? ?a $? pattern match instead. This requires the use of slot-delete$
which requires two message passes to occur per rule fire to delete the target
element within the given IndirectClaims and Claims. This increase in the number
of message passes is a faster alternative to the modify-instance route.