Broke out the producers/consumers merging into their own stages
I also removed the merging of producers and consumers after the Analysis stage
because the extra sets of merging were slowing things down. The Acts of merging
and injecting consumers, producers, and such have been moved into their own
stages. This not only reduces the work load per stage but also makes sure that
all corresponding facts have been emitted before operating on them.

Ironically, the slowest part of the wavefront scheduling pass is the act of
merging producers.

I added stages for
1) Injecting producers and consumers,
2) Setting call barriers
3) Merging producers and consumers
4) Settifying producers, consumers, and non-local dependencies

I also rewrote the injection of producers and consumers so that the checks for
existence are performed during merging instead of through rules.