Modified the act of producer injection such that local dependencies are now set in MarkLockDependencies
This change not only greatly improves the running time of the optimization but
also exposes the fact that existence checks performed by
ExtendedInjectProducers for local dependencies were overkill. At most the
number of elements within the local dependencies set will be the total number
of operands for the target element. Duplicate entries in this case mean that
the same dependency is used multiple times. It is safer to have the settify
commands take care of the uniqueness requirement then the injection rules