Undid the individual local dependency updating and added an interleaved inject rule
There is now a rule that will inject both producers and consumers if both exist
for a given instruction. If both do not exist then the previous Inject rules
take over.

I also modified MergeConsumers and MergeProducers so that the check for
inequality is done within the second conditional expression instead of a third
one. This has decreased the running time of the optimization from 1.25 seconds
for flops-no-opt to 1.18 seconds.