Rewrote how uses are replaced in CLIPS
It turns out that I was not checking to make sure that the target instruction
to have operands replaced in even contained the original target. This reduction
of the search from a brute force was easy to do by checking to see if the
target original operand exists in the instructions operands. It does require a
ton of messages being sent to get the list of operands but even with the
increase in messages the code is significantly faster.