Deleted two rules dealing with MergingSlice analysis facts
These two rules were MultiSingle merge rules. The problem is that these rules
were complex enough that generating the agenda on a change stage was causing
some functions to take nearly have a minute to schedule. Removing these two
rules have yielded a significant performance improvement.

The reason for the performance degredation has to do with what I'm matching.
Let's use the first deleted rule MergeSliceAnalysisFacts-SingleMulti to
describe the performance issues

(defrule MergeSliceAnalysisFacts-SingleMulti
         (Stage WavefrontSchedule $?)
         (Substage Analyze $?)
         ?f0 <- (Analyze slice ?s0 for ?e and cpv ?cpv)
         ?f1 <- (Analyze in ?e using cpv ?cpv and slices $?z)
         (test (not (member$ ?s0 ?z)))
         ;(test (eq FALSE (member$ ?s0 $?z)))
         =>
         (retract ?f0 ?f1)
         (assert (Analyze in ?e using cpv ?cpv and slices $?z ?s0)))

Alright, while the rule's condition is pretty simple the problem is that I'm
actually comparing two fact pools that can become quite massive. So if there
are roughly 3000 singular facts and 3000 plural facts then it's necessary to
take facts from each pool and compare them. Not only that but also make sure
that ?s0 is not a member of $?z. The problem is that this is a linear operation
so not only do we do a linear existence operation but also comparing two
massive fact pools. Combined with whatever the running time of the agenda
generator is it's possible to really slow the scheduler down.

The other side of the condition where ?s0 is a subset of $?z it doesn't really
matter because the existence check necessary as well.

Removing these two rules should speed things up greatly.

This issue falls under the same idea as trying to merge more than two facts
together at the same time.