Added a new rule to handle local dependency analysis of call instructions
Call instructions have the function being called as one of the arguments. This
can cause a problem when the name of the function being called is the same name
as the register storing the result of the call

%puts = call void @puts(.....)

LLVM provides the argument operands as a separate field. This knowledge was
always provided by LibExpertSystem. Thus it was easy to just add a new rule to
iterate over that